OPHAKTKEH-133: moved creation method in ClerkEmailService
