+++ src/test/java/fi/oph/akt/service/ClerkTranslatorServiceTest.java
@@ -5,7 +5,6 @@ import fi.oph.akt.api.dto.clerk.ClerkLanguagePairDTO;
 import fi.oph.akt.api.dto.clerk.ClerkTranslatorAuthorisationDTO;
 import fi.oph.akt.api.dto.clerk.ClerkTranslatorDTO;
 import fi.oph.akt.api.dto.clerk.ClerkTranslatorResponseDTO;
-import fi.oph.akt.api.dto.clerk.InformalEmailRequestDTO;
 import fi.oph.akt.model.Authorisation;
 import fi.oph.akt.model.AuthorisationBasis;
 import fi.oph.akt.model.AuthorisationTerm;
@@ -17,16 +16,11 @@ import fi.oph.akt.repository.AuthorisationRepository;
 import fi.oph.akt.repository.AuthorisationTermRepository;
 import fi.oph.akt.repository.LanguagePairRepository;
 import fi.oph.akt.repository.TranslatorRepository;
-import fi.oph.akt.service.email.EmailData;
-import fi.oph.akt.service.email.EmailService;
 import org.junit.jupiter.api.BeforeEach;
 import org.junit.jupiter.api.Test;
-import org.mockito.ArgumentCaptor;
-import org.mockito.Captor;
 import org.springframework.beans.factory.annotation.Autowired;
 import org.springframework.boot.test.autoconfigure.orm.jpa.DataJpaTest;
 import org.springframework.boot.test.autoconfigure.orm.jpa.TestEntityManager;
-import org.springframework.boot.test.mock.mockito.MockBean;
 import org.springframework.context.annotation.Import;
 
 import javax.annotation.Resource;
@@ -37,11 +31,7 @@ import java.util.stream.IntStream;
 
 import static org.junit.jupiter.api.Assertions.assertEquals;
 import static org.junit.jupiter.api.Assertions.assertNull;
-import static org.junit.jupiter.api.Assertions.assertThrows;
 import static org.junit.jupiter.api.Assertions.assertTrue;
-import static org.mockito.ArgumentMatchers.any;
-import static org.mockito.Mockito.times;
-import static org.mockito.Mockito.verify;
 
 @DataJpaTest
 @Import({ OnrServiceMock.class })
@@ -55,9 +45,6 @@ class ClerkTranslatorServiceTest {
 	@Resource
 	private AuthorisationTermRepository authorisationTermRepository;
 
-	@MockBean
-	private EmailService emailService;
-
 	@Resource
 	private LanguagePairRepository languagePairRepository;
 
@@ -70,13 +57,10 @@ class ClerkTranslatorServiceTest {
 	@Resource
 	private TestEntityManager entityManager;
 
-	@Captor
-	private ArgumentCaptor<EmailData> emailDataCaptor;
-
 	@BeforeEach
 	public void setup() {
 		clerkTranslatorService = new ClerkTranslatorService(authorisationRepository, authorisationTermRepository,
-				emailService, languagePairRepository, translatorRepository, onrServiceMock);
+				languagePairRepository, translatorRepository, onrServiceMock);
 	}
 
 	@Test
@@ -398,76 +382,6 @@ class ClerkTranslatorServiceTest {
 
 		assertEquals(term3BeginDate, authorisationDTO.term().beginDate());
 		assertEquals(term3EndDate, authorisationDTO.term().endDate());
-	}
-
-	@Test
-	public void createInformalEmailsShouldSaveEmailsToGivenTranslators() {
-		final MeetingDate meetingDate = Factory.meetingDate();
-		entityManager.persist(meetingDate);
-
-		IntStream.range(0, 3).forEach(n -> {
-			final Translator translator = Factory.translator();
-			final Authorisation authorisation = Factory.authorisation(translator, meetingDate);
-			final LanguagePair languagePair = Factory.languagePair(authorisation);
-			final AuthorisationTerm authorisationTerm = Factory.authorisationTerm(authorisation);
-
-			entityManager.persist(translator);
-			entityManager.persist(authorisation);
-			entityManager.persist(languagePair);
-			entityManager.persist(authorisationTerm);
-		});
-
-		List<Long> translatorIds = translatorRepository.findAll().stream().map(Translator::getId).toList();
-
-		InformalEmailRequestDTO emailRequestDTO = InformalEmailRequestDTO.builder().translatorIds(translatorIds)
-				.subject("otsikko").body("viesti").build();
-
-		clerkTranslatorService.createInformalEmails(emailRequestDTO);
-
-		verify(emailService, times(3)).saveEmail(any(), emailDataCaptor.capture());
-
-		List<EmailData> emailDatas = emailDataCaptor.getAllValues();
-
-		assertEquals(3, emailDatas.size());
-
-		emailDatas.forEach(emailData -> {
-			assertEquals("AKT", emailData.sender());
-			assertEquals("otsikko", emailData.subject());
-			assertEquals("viesti", emailData.body());
-		});
-	}
-
-	@Test
-	public void createInformalEmailsShouldSaveEmailToGivenTranslatorsWithDuplicateTranslatorIds() {
-		final MeetingDate meetingDate = Factory.meetingDate();
-		final Translator translator = Factory.translator();
-		final Authorisation authorisation = Factory.authorisation(translator, meetingDate);
-		final LanguagePair languagePair = Factory.languagePair(authorisation);
-		final AuthorisationTerm authorisationTerm = Factory.authorisationTerm(authorisation);
-
-		entityManager.persist(meetingDate);
-		entityManager.persist(translator);
-		entityManager.persist(authorisation);
-		entityManager.persist(languagePair);
-		entityManager.persist(authorisationTerm);
-
-		Long tId = translatorRepository.findAll().get(0).getId();
-
-		InformalEmailRequestDTO emailRequestDTO = InformalEmailRequestDTO.builder().translatorIds(List.of(tId, tId))
-				.subject("otsikko").body("viesti").build();
-
-		clerkTranslatorService.createInformalEmails(emailRequestDTO);
-
-		verify(emailService).saveEmail(any(), emailDataCaptor.capture());
-	}
-
-	@Test
-	public void createInformalEmailsShouldThrowIllegalArgumentExceptionForNonExistingTranslatorIds() {
-		InformalEmailRequestDTO emailRequestDTO = InformalEmailRequestDTO.builder().translatorIds(List.of(1L))
-				.subject("otsikko").body("viesti").build();
-
-		assertThrows(IllegalArgumentException.class,
-				() -> clerkTranslatorService.createInformalEmails(emailRequestDTO));
 	}
 
 }
